+++ MVHS.js
@@ -4,7 +4,7 @@
 //
 "use strict";
 
-const MVHSVersion = "2.1.2";
+const MVHSVersion = "2.2.1";
 
 // Revision History
 //
@@ -37,6 +37,9 @@ const MVHSVersion = "2.1.2";
 //
 //  2.2.0   01/24/2021  Added support for cookies and the ability to change class
 //                      choices
+//
+//  2.2.1   01/26/2021  Added 'teachersFromCookie' and 'roomsFromCookie' as unused
+//                      arrays of data that may be implemented in the future
 
 // =+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=
 // Copyright 2020 Mike Uhler and Jonathan Uhler
@@ -134,7 +137,8 @@ class SchoolYearDefinitions {
     this.lastPeriod = _lastPeriod_k;
 
     // Export the student name
-    this.studentName = "Jonathan Uhler";
+    this.studentName = "Student Name"; // This is currently unused, although it 
+                                       // may be put into the cookie at some point
 
     // Define the types of a day
     const _dayTypeSchoolDay_k = "School Day";
@@ -163,25 +167,29 @@ class SchoolYearDefinitions {
     // Read in cookie data
     try {
       var cookieData = document.cookie
-      var periodsFromCookie = cookieData.split("=")
-      periodsFromCookie = periodsFromCookie[1].split(",")
+      var cookieSections = cookieData.split("=")
+      var periodsFromCookie = cookieSections[1].split(",")
+      var roomsFromCookie = cookieSections[3].split(",")
+      var teachersFromCookie = cookieSections[5].split(",")
     }
     catch {
       periodsFromCookie = ["None", "None", "None", "None", "None", "None", "None"]
+      roomsFromCookie = ["None", "None", "None", "None", "None", "None", "None"]
+      teachersFromCookie = ["None", "None", "None", "None", "None", "None", "None"]
     }
     
 
     // Class information array, indexed by period
     const _classInfoArray = [
       // period     class                Room        Teacher
-      {p: 0,    c:  null,                r: "",    t: ""            },
-      {p: 1,    c: periodsFromCookie[0], r: "113", t: "Kim Rogers"  },
-      {p: 2,    c: periodsFromCookie[1], r: "Gym", t: "Williams"    },
-      {p: 3,    c: periodsFromCookie[2], r: "602", t: "Cardenas"    },
-      {p: 4,    c: periodsFromCookie[3], r: "215", t: "Engel-Hall"  },
-      {p: 5,    c: periodsFromCookie[4], r: "412", t: "Smith"       },
-      {p: 6,    c: periodsFromCookie[5], r: "",    t: ""            },
-      {p: 7,    c: periodsFromCookie[6], r: "514", t: "Dilloughery" }
+      {p: 0,    c:  null,                r: "",                 t: ""                    },
+      {p: 1,    c: periodsFromCookie[0], r: roomsFromCookie[0], t: teachersFromCookie[0] },
+      {p: 2,    c: periodsFromCookie[1], r: roomsFromCookie[1], t: teachersFromCookie[1] },
+      {p: 3,    c: periodsFromCookie[2], r: roomsFromCookie[2], t: teachersFromCookie[2] },
+      {p: 4,    c: periodsFromCookie[3], r: roomsFromCookie[3], t: teachersFromCookie[3] },
+      {p: 5,    c: periodsFromCookie[4], r: roomsFromCookie[4], t: teachersFromCookie[4] },
+      {p: 6,    c: periodsFromCookie[5], r: roomsFromCookie[5], t: teachersFromCookie[5] },
+      {p: 7,    c: periodsFromCookie[6], r: roomsFromCookie[6], t: teachersFromCookie[6] }
     ];
 
     // -------------------------------------------------------------------------